requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
# Pin bcrypt to avoid incompatibility with passlib expecting __about__
bcrypt<4.0.0,>=3.2.2
email-validator==2.2.0
python-multipart==0.0.9
